refactor: Refactor & add klc generation to cli
